AoC2017.Tests/Day18Test.cs
using System;
using System.Collections.Generic;
using System.Linq;
using Xunit;

namespace AoC2017.Tests
{
    public class Day18Test
    {
        [Theory]
        [InlineData(@"set a 1
add a 2
mul a a
mod a 5
snd a
set a 0
rcv a
jgz a -1
set a 1
jgz a -2", 4)]
        public void Part1(string input, int output)
        {
            var result = Day18.Part1(input);
            Assert.True(result == output, $"Expected: {output}, Received: {result}");
        }

        [Theory]
        [InlineData("", 0)]
        public void Part2(string input, int output)
        {
            var result = Day18.Part2(input);
            Assert.True(result == output, $"Expected: {output}, Received: {result}");
        }
    }
}
